Refactor id generation
